Add CSS reset and Three.js setup
This commit introduces a CSS reset to normalize styles across different browsers by setting body margin to zero in style.css. A link to this stylesheet has been added to index.html. Also, it introduces basic setup for a three.js scene in main.js which includes the creation of a scene, camera, renderer, geometry, material, a mesh (cube) and an animate function for animation. This sets up the groundwork for the inclusion of WebGL 3D animations in our project.