Refactor script loading and rename onClick function
Improved loading efficiency by moving the main.js script tag into head with 'defer' attribute in index.html. This ensures that the script is fetched in parallel with the HTML parsing and executed only after HTML parsing is finished, reducing render-blocking.

In main.js, renamed `onClick` function to `clickHandler` for better clarity and added a console.log for mouse click and stone position to improve debugging and tracking user interaction.