Implemented two-player functionality in 3D game
A new feature of two-player functionality has been added to the 3D game. Besides, the game board is now initialized as a 2D array. A plane geometry is now created for raycaster intersection. A clickHandler function has also been introduced that calculates the coordinates of the objects intersected when clicked and changes the color of the stones based on the current player. Commented out lines of code preceding the feature have been kept for future reference.