"Refactor game code for better maintainability and add player turn information"
This commit refactors the main.js file in order to improve code maintainability and readability. The changes include extracting functions in a single blob of code, which makes it easier to understand its function. Also, constant values are introduced to replace hard-coded values to enhance the clarity of their use and simplify adjustments. This refactoring doesn't change any game logic. Moreover, an additional div element for displaying the current player's turn has been added. This allows users to keep track of whose turn it is, thus improving the user experience.