Add tracking and removal for stones in scene
Added a new array, stonesInScene, to keep track of all the stones present in the game scene. This array is used to efficiently remove all the stones from the scene when clearing the game board. The approach improves the clearing process by avoiding a loop through the entire game board which was previously dependent upon the game board to remove stones from the scene.