Added win condition check to game logic
A new constant WINNING_NUMBER_OF_STONES was introduced to define what conditions need to be met for a player to have won the game. Subsequently, a new function checkWin was added to scan the game board for any sequences of five or more stones of the same color. This check is performed after each placement of a stone. Changes were also made to the handleMouseUp function to incorporate this new check. Furthermore, an alteration was made to the way the game board state is stored - using color strings instead of stone objects for better performance during win check.