Add score tracking and improve player turn visual
Made few modifications to the code for a clearer, more interactive gameplay experience. The CSS styling was changed to better distinguish which player's turn is currently active. A `player-info` class has been introduced with specific color designations for black and white players.

A significant improvement which has been introduced is score tracking. Now, the game keeps track of scores for both players. The current score for both the black and the white player is displayed on a new scoreboard. This was accomplished by creating a new `score-board` div element in the Javascript file. It is updated in the `countConsecutiveSameColorStonesAndUpdateScore` function each time a player makes a winning move.

The modification also included minor optimization like removal of unnecessary comments and changes in player label's position.