Add replay functionality to game interface
A new replay button and accompanying functionality was added to allow users to replay games. This included CSS changes for the new replay button, as well as updates to the JavaScript file to handle replay actions. Gameplay data is now stored in the `stonesInScene` array so it can be reused during a replay. Additionally, a 'replay' state was added to the game state to control replay actions. The drawStone function was refactored to accommodate the new replay functionality.